requirements.txt
colorlog>=6.9.0
mypy>=1.13.0
pre-commit>=4.2.0
ruff>=0.4.0

# For tests
pytest==8.4.2
pytest-asyncio==1.2.0
pytest_homeassistant_custom_component==0.13.286
# aiodns used by the HA test stack is incompatible with pycares 5.x on Python 3.13.
pycares<5
